attempt to reduce timings

+++ src/functions.cpp
@@ -9,7 +9,7 @@ double distance(double x0, double y0, double x1, double y1) {
 void import_rij(Rcpp::List& rij_list, std::vector<arma::sp_mat>& rij) {
   arma::sp_mat m;
   rij.resize(rij_list.size());
-  for (std::size_t i = 0; i < rij_list.size(); ++i) {
+  for (std::size_t i = 0; i < static_cast<std::size_t>(rij_list.size()); ++i) {
     m = Rcpp::as<arma::sp_mat>(rij_list[i]);
     rij[i] = m.t();
   }
